fix: flaky test
